split out profile_selector component
